Editar y Borrar con caja de diálogo

+++ app/src/main/java/com/example/listadealumnos/AlumnoAdapter.kt
@@ -1,6 +1,7 @@
 package com.example.listadealumnos
 
 import Alumno
+import android.app.Activity
 import android.content.Intent
 import android.util.Log
 import android.view.LayoutInflater
@@ -10,6 +11,9 @@ import android.widget.ImageView
 import android.widget.LinearLayout
 import android.widget.TextView
 import android.widget.Toast
+import androidx.activity.result.ActivityResultLauncher
+import androidx.activity.result.contract.ActivityResultContracts
+import androidx.appcompat.app.AlertDialog
 import androidx.appcompat.widget.PopupMenu
 import androidx.recyclerview.widget.RecyclerView
 import com.squareup.picasso.Picasso
@@ -21,7 +25,8 @@ class CustomViewHolder(itemView: View) : RecyclerView.ViewHolder(itemView) {
     val containerOpciones: TextView = itemView.findViewById(R.id.textViewOptions)
     val containerAlumno: LinearLayout = itemView.findViewById(R.id.alumnoLayout)
 }
-class AlumnoAdapter(private val data: MutableList<Alumno>, private val dbAlumno: DBHelperAlumno) : RecyclerView.Adapter<CustomViewHolder>() {
+class AlumnoAdapter(private val data: MutableList<Alumno>, private val dbAlumno: DBHelperAlumno, val resultLauncher: ActivityResultLauncher<Intent?>) : RecyclerView.Adapter<CustomViewHolder>() {
+
     override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): CustomViewHolder {
         val itemView = LayoutInflater.from(parent.context).inflate(R.layout.alumno_layout, parent, false)
         return CustomViewHolder(itemView)
@@ -59,30 +64,66 @@ class AlumnoAdapter(private val data: MutableList<Alumno>, private val dbAlumno:
     private fun showOptionsMenu(view: View, idAlumno: Int) {
         val popupMenu = PopupMenu(view.context, view)
         val inflater = popupMenu.menuInflater
+        var alumno: Alumno? = null;
+        for(i in data.indices) {
+            val tempAlumno = data[i]
+            if (tempAlumno.id == idAlumno) {
+                alumno = tempAlumno
+                break
+            }
+        }
+        if(alumno == null) {
+            Toast.makeText(view.context, "No se pudo encontrar el alumno", Toast.LENGTH_SHORT).show()
+            return;
+        }
         inflater.inflate(R.menu.menu_main, popupMenu.menu)
 
         popupMenu.setOnMenuItemClickListener { item ->
             when (item.itemId) {
                 R.id.borrar -> {
-                    val db = dbAlumno.writableDatabase
-                    val condicion = "id = ?"
-                    val argumentos = arrayOf(idAlumno.toString())
-                    db.delete("alumnos", condicion, argumentos)
-                    db.close()
 
-                    for(i in data.indices) {
-                        val alumno = data[i]
-                        if (alumno.id == idAlumno) {
-                            data.remove(alumno)
-                            this.notifyItemRemoved(i)
-                            break
+                    val builder = AlertDialog.Builder(view.context)
+                    builder.setTitle("Borrar Alumno")
+                    builder.setMessage("¿Deseas borrar el alumno '${alumno.nombre}'?")
+
+                    // Positive button (Yes)
+                    builder.setPositiveButton("Si") { dialog, which ->
+                        val db = dbAlumno.writableDatabase
+                        val condicion = "id = ?"
+                        val argumentos = arrayOf(idAlumno.toString())
+                        db.delete("alumnos", condicion, argumentos)
+                        db.close()
+
+                        for(i in data.indices) {
+                            val alumno = data[i]
+                            if (alumno.id == idAlumno) {
+                                data.remove(alumno)
+                                this.notifyItemRemoved(i)
+                                break
+                            }
                         }
                     }
+
+                    // Negative button (No)
+                    builder.setNegativeButton("No") { dialog, which -> }
+
+                    // Create and show the dialog
+                    val dialog: AlertDialog = builder.create()
+                    dialog.show()
                     true
                 }
                 R.id.editar -> {
-                    // Handle the "Delete" action here
-                    TODO();
+                    val intent = Intent(view.context, EditarAlumno::class.java)
+
+                    // Add any data you want to pass to the ViewAlumno activity (optional)
+                    intent.putExtra("id", alumno.id)
+                    intent.putExtra("nombre", alumno.nombre)
+                    intent.putExtra("numCuenta", alumno.cuenta)
+                    intent.putExtra("email", alumno.correo)
+                    intent.putExtra("imagen", alumno.imagen)
+
+                    // Start the activity
+                    resultLauncher.launch(intent)
                     true
                 }
                 else -> false
